Move FilePermissions into Bolt\Filesystem namespace
